Rename to diplomat_c_runtime.hpp
