+++ tool/templates/cpp2/runtime.hpp
@@ -11,7 +11,7 @@
 #include<span>
 #endif
 
-#include "diplomat_runtime.h"
+#include "diplomat_c_runtime.hpp"
 
 namespace diplomat {
